agent,billing: move pkg/billing
Move event definitions to pkg/reporting.
Move hostname helper to the pkg/agent/billing by place of usage.
Move schema file billing-v1.yaml to s3 e2e test by place of single
usage.
