+++ pkg/agent/billing/clients.go
@@ -11,7 +11,6 @@ import (
 	"github.com/lithammer/shortuuid"
 	"go.uber.org/zap"
 
-	"github.com/neondatabase/autoscaling/pkg/billing"
 	"github.com/neondatabase/autoscaling/pkg/reporting"
 )
 
@@ -38,7 +37,7 @@ type HTTPClientConfig struct {
 	URL string `json:"url"`
 }
 
-type billingClient = reporting.Client[*billing.IncrementalEvent]
+type billingClient = reporting.Client[*reporting.IncrementalEvent]
 
 func createClients(ctx context.Context, logger *zap.Logger, cfg ClientsConfig) ([]billingClient, error) {
 	var clients []billingClient
@@ -92,9 +91,9 @@ func createClients(ctx context.Context, logger *zap.Logger, cfg ClientsConfig) (
 	return clients, nil
 }
 
-func jsonMarshalEvents(events []*billing.IncrementalEvent) ([]byte, reporting.SimplifiableError) {
+func jsonMarshalEvents(events []*reporting.IncrementalEvent) ([]byte, reporting.SimplifiableError) {
 	obj := struct {
-		Events []*billing.IncrementalEvent `json:"events"`
+		Events []*reporting.IncrementalEvent `json:"events"`
 	}{Events: events}
 
 	return reporting.JSONMarshalBatch(&obj)
