+++ pkg/reporting/events.go
@@ -1,6 +1,7 @@
-package billing
+package reporting
 
 import (
+	"fmt"
 	"time"
 )
 
@@ -68,4 +69,21 @@ func (e *IncrementalEvent) setType() {
 // getIdempotencyKey implements eventMethods
 func (e *IncrementalEvent) getIdempotencyKey() *string {
 	return &e.IdempotencyKey
+}
+
+// Enrich sets the event's Type and IdempotencyKey fields, so that users of this API don't need to
+// manually set them
+func Enrich[E Event](now time.Time, hostname string, countInBatch, batchSize int, event E) E {
+	event.setType()
+
+	// RFC3339 with microsecond precision. Possible to get collisions with millis, nanos are extra.
+	// And everything's in UTC, so there's no sense including the offset.
+	formattedTime := now.In(time.UTC).Format("2006-01-02T15:04:05.999999Z")
+
+	key := event.getIdempotencyKey()
+	if *key == "" {
+		*key = fmt.Sprintf("%s-%s-%d/%d", formattedTime, hostname, countInBatch, batchSize)
+	}
+
+	return event
 }
